Optimized Drag n Drop for both texture and JSON
Also, given a JSON file: the entity factory will now check the type within the json file instead of trying to construct it from the filename which is error prone. If the json file does not have a "type" member, it will resort to this behavior and a warning will be logged.